requirements.txt
google-cloud-error-reporting==1.5.0